requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
psycopg2==2.8.4

flake8==4.0.1
